Assign labels as feeds are imported.
Switches listparser for Beautiful Soup as this gives us the tags more
easily.

Also removes ZipFile parsing. Although it's convenient to just throw
a Google Reader takeout export at it, not everyone will be doing this
directly and it's no great hardship to unzip and find the XML yourself.

Fixes #13.